LP#1646166 Hatch Windows Installer
Code for building a Windows installer executable for Hatch.

The code makes no attempt to install a JRE.  It only verifies that a JRE
version 1.8 or above is present.  It also does not add the browser
extension to the browser, it only installs the files needed for the
extension to work once added to the browser.

See installer/windows/README.adoc for instructions on building the
Windows installer.

Signed-off-by: Kyle Huckins <khuckins@catalystdevworks.com>
Signed-off-by: Bill Erickson <berickxx@gmail.com>